requirements.txt
python-dotenv==1.0.1
discord.py==2.5.2
Flask==3.1.0
selenium==4.30.0
undetected-chromedriver==3.5.5